core/libs/execution/pyproject.toml
[project]
name = "agentarea-execution"
version = "0.1.0"
description = "AgentArea execution library using Google ADK (Agent Development Kit)"
requires-python = ">=3.9"
dependencies = [
    # Core dependencies
    "temporalio>=1.0.0",
    "pydantic>=2.0.0",
    
    # Google ADK - Agent Development Kit
    "google-adk>=0.1.0",
    
    # HTTP client
    "httpx>=0.24.0",
    
    # For testing only - remove these for production
    # "agentarea-common",
    # "agentarea-agents",
    # "agentarea-mcp",
    # "agentarea-llm",
    # "agentarea-tasks",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentarea_execution"] 